Rewrite for better chaining mutations
This rewrites the internal implementation to allow proper chaining like
Promises do. It now returns new Confirmer objects instead of mutating
the original Confirmer like it used to. It also adds a low level
Confirmer.resolve method meant to over step the initializer function
and directly apply the internal promise. It is meant to support the
creation of a new Confirmer in place. Used internally it is exposed for
advanced use cases by the consumers.